Move loader functions into seperate file
