+++ lib/CMakeLists.txt
@@ -40,6 +40,7 @@ set(LIBIRVE_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/gdbserver.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/gdbserver.h
     ${CMAKE_CURRENT_SOURCE_DIR}/irve_public_api.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/loaders.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/logging.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/logging.h
     ${CMAKE_CURRENT_SOURCE_DIR}/memory.cpp
